Move ws related helpers to api/ws/ folder
