+++ framework/src/modules/chain/logic/peers.js
@@ -17,8 +17,8 @@
 const _ = require('lodash');
 const failureCodes = require('../api/ws/rpc/failure_codes.js');
 const Peer = require('../logic/peer.js');
-const PeersManager = require('../helpers/peers_manager.js');
-const BanManager = require('../helpers/ban_manager.js');
+const PeersManager = require('../api/ws/peers_manager.js');
+const BanManager = require('../api/ws/ban_manager.js');
 
 // Private fields
 let self;
